* Extracted zipper functions from http handlers * Removed App abstraction from the cache probing
