+++ pkg/app/zipper/http_handlers.go
@@ -16,7 +16,6 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"sort"
 	"strconv"
 	"time"
 
@@ -81,62 +80,15 @@ func (app *App) findHandler(w http.ResponseWriter, req *http.Request, ms *Promet
 		kv.String("graphite.format", format),
 		kv.String("graphite.target", originalQuery),
 	)
-	request := types.NewFindRequest(originalQuery)
-	bs := app.filterBackendByTopLevelDomain([]string{originalQuery})
-	var filteredByPathCache bool
-	bs, filteredByPathCache = backend.Filter(bs, []string{originalQuery})
-	if filteredByPathCache {
-		ms.PathCacheFilteredRequests.Inc()
-	}
-	metrics, errs := backend.Finds(ctx, bs, request, app.Metrics.FindOutDuration)
-	err := errorsFanIn(errs, len(bs))
 
-	if ctx.Err() != nil {
-		// context was cancelled even if some of the requests succeeded
-		app.Metrics.RequestCancel.WithLabelValues(
-			"find", ctx.Err().Error(),
-		).Inc()
-	}
-
+	metrics, err := Find(app, ctx, originalQuery, &span, ms, logger)
 	if err != nil {
-		span.SetAttribute("error", true)
-		span.SetAttribute("error.message", err.Error())
-		var notFound types.ErrNotFound
-		if errors.As(err, &notFound) {
-			// graphite-web 0.9.12 needs to get a 200 OK response with an empty
-			// body to be happy with its life, so we can't 404 a /metrics/find
-			// request that finds nothing. We are however interested in knowing
-			// that we found nothing on the monitoring side, so we claim we
-			// returned a 404 code to Prometheus.
-
-			app.Metrics.FindNotFound.Inc()
-			logger.Info("not found",
-				zap.Error(err))
-			// TODO (grzkv) Should we return here?
-		} else {
-			code := http.StatusInternalServerError
-			logger.Error("find failed",
-				zap.Int("http_code", code),
-				zap.Duration("runtime_seconds", time.Since(t0)),
-				zap.Error(err),
-			)
-			http.Error(w, err.Error(), code)
-			app.Metrics.Responses.WithLabelValues(strconv.Itoa(code), "find").Inc()
-			return
-		}
+		code := http.StatusInternalServerError
+		logger.Error("find failed", zap.Int("http_code", code), zap.Duration("runtime_seconds", time.Since(t0)), zap.Error(err))
+		http.Error(w, err.Error(), code)
+		ms.Responses.WithLabelValues(strconv.Itoa(code), "find").Inc()
+		return
 	}
-
-	sort.Slice(metrics.Matches, func(i, j int) bool {
-		if metrics.Matches[i].Path < metrics.Matches[j].Path {
-			return true
-		}
-		if metrics.Matches[i].Path > metrics.Matches[j].Path {
-			return false
-		}
-		return metrics.Matches[i].Path < metrics.Matches[j].Path
-	})
-
-	span.SetAttribute("graphite.total_metric_count", len(metrics.Matches))
 
 	var contentType string
 	var blob []byte
@@ -160,12 +112,8 @@ func (app *App) findHandler(w http.ResponseWriter, req *http.Request, ms *Promet
 
 	if err != nil {
 		http.Error(w, "error marshaling data", http.StatusInternalServerError)
-		logger.Error("find failed",
-			zap.Int("http_code", http.StatusInternalServerError),
-			zap.String("reason", "error marshaling data"),
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.Error(err),
-		)
+		logger.Error("find failed", zap.Int("http_code", http.StatusInternalServerError), zap.String("reason", "error marshaling data"),
+			zap.Duration("runtime_seconds", time.Since(t0)), zap.Error(err))
 		app.Metrics.Responses.WithLabelValues(strconv.Itoa(http.StatusInternalServerError), "find").Inc()
 		span.SetAttribute("error", true)
 		span.SetAttribute("error.message", err.Error())
@@ -177,26 +125,19 @@ func (app *App) findHandler(w http.ResponseWriter, req *http.Request, ms *Promet
 
 	if writeErr != nil {
 		app.Metrics.Responses.WithLabelValues(strconv.Itoa(499), "find").Inc()
-		logger.Warn("error writing the response",
-			zap.Int("http_code", 499),
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.Error(writeErr),
-		)
+		logger.Warn("error writing the response", zap.Int("http_code", 499), zap.Duration("runtime_seconds", time.Since(t0)), zap.Error(writeErr))
 		return
 	}
 
 	app.Metrics.Responses.WithLabelValues(strconv.Itoa(http.StatusOK), "find").Inc()
-	logger.Info("request served",
-		zap.Int("http_code", http.StatusOK),
-		zap.Duration("runtime_seconds", time.Since(t0)),
-	)
+	logger.Info("request served", zap.Int("http_code", http.StatusOK), zap.Duration("runtime_seconds", time.Since(t0)))
 }
 
 func (app *App) renderHandler(w http.ResponseWriter, req *http.Request, ms *PrometheusMetrics, logger *zap.Logger) {
 	t0 := time.Now()
 
-	tExp := prometheus.NewTimer(app.Metrics.RenderDurationExp)
-	defer tExp.ObserveDuration()
+	t := prometheus.NewTimer(app.Metrics.RenderDurationExp)
+	defer t.ObserveDuration()
 
 	memoryUsage := 0
 
@@ -205,9 +146,7 @@ func (app *App) renderHandler(w http.ResponseWriter, req *http.Request, ms *Prom
 	span := trace.SpanFromContext(ctx)
 
 	if ce := logger.Check(zap.DebugLevel, "got render request"); ce != nil {
-		ce.Write(
-			zap.String("request", req.URL.RequestURI()),
-		)
+		ce.Write(zap.String("request", req.URL.RequestURI()))
 	}
 
 	app.Metrics.Requests.Inc()
@@ -233,14 +172,9 @@ func (app *App) renderHandler(w http.ResponseWriter, req *http.Request, ms *Prom
 
 	target := req.FormValue("target")
 	format := req.FormValue("format")
-	logger = logger.With(
-		zap.String("format", format),
-		zap.String("target", target),
-	)
-	span.SetAttributes(
-		kv.String("graphite.target", target),
-		kv.String("graphite.format", format),
-	)
+	logger = logger.With( zap.String("format", format), zap.String("target", target))
+	span.SetAttributes( kv.String("graphite.target", target), kv.String("graphite.format", format))
+
 	from, err := strconv.ParseInt(req.FormValue("from"), 10, 64)
 	if err != nil {
 		http.Error(w, "from is not a integer", http.StatusBadRequest)
@@ -273,10 +207,7 @@ func (app *App) renderHandler(w http.ResponseWriter, req *http.Request, ms *Prom
 		return
 	}
 
-	span.SetAttributes(
-		kv.Int64("graphite.from", from),
-		kv.Int64("graphite.until", until),
-	)
+	span.SetAttributes(kv.Int64("graphite.from", from), kv.Int64("graphite.until", until))
 
 	if target == "" {
 		http.Error(w, "empty target", http.StatusBadRequest)
@@ -292,26 +223,12 @@ func (app *App) renderHandler(w http.ResponseWriter, req *http.Request, ms *Prom
 		return
 	}
 
-	request := types.NewRenderRequest([]string{target}, int32(from), int32(until))
-	request.Trace.OutDuration = app.Metrics.RenderOutDurationExp
-	bs := app.filterBackendByTopLevelDomain(request.Targets)
-	var filteredByPathCache bool
-	bs, filteredByPathCache = backend.Filter(bs, request.Targets)
-	if filteredByPathCache {
-		ms.PathCacheFilteredRequests.Inc()
-	}
-	metrics, stats, errs := backend.Renders(ctx, bs, request, app.Config.RenderReplicaMismatchConfig, logger)
-	app.Metrics.Renders.Add(float64(stats.DataPointCount))
-	app.Metrics.RenderMismatches.Add(float64(stats.MismatchCount))
-	app.Metrics.RenderFixedMismatches.Add(float64(stats.FixedMismatchCount))
-	err = errorsFanIn(errs, len(bs))
-	span.SetAttribute("graphite.metrics", len(metrics))
+	metrics, err := Render(app, ctx, target, from, until, ms, logger)
 
+	span.SetAttribute("graphite.metrics", len(metrics))
 	if ctx.Err() != nil {
 		// context was cancelled even if some of the requests succeeded
-		app.Metrics.RequestCancel.WithLabelValues(
-			"render", ctx.Err().Error(),
-		).Inc()
+		app.Metrics.RequestCancel.WithLabelValues("render", ctx.Err().Error()).Inc()
 		span.SetAttribute("error", true)
 		span.SetAttribute("error.message", ctx.Err().Error())
 	}
@@ -327,11 +244,7 @@ func (app *App) renderHandler(w http.ResponseWriter, req *http.Request, ms *Prom
 
 		http.Error(w, msg, code)
 		fields := []zap.Field{
-			zap.Int("memory_usage_bytes", memoryUsage),
-			zap.Error(err),
-			zap.Int("http_code", code),
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.Int64s("trace", request.Trace.Report()),
+			zap.Int("memory_usage_bytes", memoryUsage), zap.Error(err), zap.Int("http_code", code), zap.Duration("runtime_seconds", time.Since(t0)),
 		}
 		if code == http.StatusNotFound {
 			logger.Info("request failed", fields...)
@@ -369,7 +282,6 @@ func (app *App) renderHandler(w http.ResponseWriter, req *http.Request, ms *Prom
 			zap.Duration("runtime_seconds", time.Since(t0)),
 			zap.Int("memory_usage_bytes", memoryUsage),
 			zap.Error(err),
-			zap.Int64s("trace", request.Trace.Report()),
 		)
 		app.Metrics.Responses.WithLabelValues(strconv.Itoa(http.StatusInternalServerError), "render").Inc()
 		span.SetAttribute("error", true)
@@ -383,25 +295,13 @@ func (app *App) renderHandler(w http.ResponseWriter, req *http.Request, ms *Prom
 
 	if writeErr != nil {
 		app.Metrics.Responses.WithLabelValues(strconv.Itoa(499), "render").Inc()
-		logger.Warn("error writing the response",
-			zap.Int("http_code", 499),
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.Error(writeErr),
-		)
+		logger.Warn("error writing the response", zap.Int("http_code", 499), zap.Duration("runtime_seconds", time.Since(t0)), zap.Error(writeErr))
 		return
 	}
 
 	app.Metrics.Responses.WithLabelValues(strconv.Itoa(http.StatusOK), "render").Inc()
-	if stats.MismatchCount > stats.FixedMismatchCount {
-		app.Metrics.RenderMismatchedResponses.Inc()
-	}
 
-	logger.Info("request served",
-		zap.Int("memory_usage_bytes", memoryUsage),
-		zap.Int("http_code", http.StatusOK),
-		zap.Duration("runtime_seconds", time.Since(t0)),
-		zap.Int64s("trace", request.Trace.Report()),
-	)
+	logger.Info("request served", zap.Int("memory_usage_bytes", memoryUsage), zap.Int("http_code", http.StatusOK), zap.Duration("runtime_seconds", time.Since(t0)))
 }
 
 func (app *App) infoHandler(w http.ResponseWriter, req *http.Request, ms *PrometheusMetrics, logger *zap.Logger) {
@@ -533,97 +433,12 @@ func (app *App) lbCheckHandler(w http.ResponseWriter, req *http.Request, ms *Pro
 	t0 := time.Now()
 
 	if ce := logger.Check(zap.DebugLevel, "loadbalancer"); ce != nil {
-		ce.Write(
-			zap.String("request", req.URL.RequestURI()),
-		)
+		ce.Write( zap.String("request", req.URL.RequestURI()))
 	}
 
 	app.Metrics.Requests.Inc()
 
 	fmt.Fprintf(w, "Ok\n")
-	logger.Info("lb request served",
-		zap.Int("http_code", http.StatusOK),
-		zap.Duration("runtime_seconds", time.Since(t0)),
-	)
-	app.Metrics.Responses.WithLabelValues(strconv.Itoa(http.StatusOK),
-		"lbcheck").Inc()
-}
-
-func (app *App) filterBackendByTopLevelDomain(targets []string) []backend.Backend {
-	targetTlds := make([]string, 0, len(targets))
-	for _, target := range targets {
-		targetTlds = append(targetTlds, getTargetTopLevelDomain(target, app.TLDPrefixes))
-	}
-
-	bs := app.filterByTopLevelDomain(app.Backends, targetTlds)
-	if len(bs) > 0 {
-		return bs
-	}
-	return app.Backends
-}
-
-func (app *App) filterByTopLevelDomain(backends []backend.Backend, targetTLDs []string) []backend.Backend {
-	bs := make([]backend.Backend, 0)
-	allTLDBackends := make([]*backend.Backend, 0)
-
-	topLevelDomainCache, _ := app.TopLevelDomainCache.Get("tlds")
-	tldCache := make(map[string][]*backend.Backend)
-	if x, ok := topLevelDomainCache.(map[string][]*backend.Backend); ok {
-		tldCache = x
-	}
-
-	if tldCache == nil {
-		return backends
-	}
-	alreadyAddedBackends := make(map[string]bool)
-	for _, target := range targetTLDs {
-		tldBackends := tldCache[target]
-		for _, backend := range tldBackends {
-			a := *backend
-			if !alreadyAddedBackends[a.GetServerAddress()] {
-				alreadyAddedBackends[a.GetServerAddress()] = true
-				allTLDBackends = append(allTLDBackends, backend)
-			}
-		}
-	}
-	for _, tldBackend := range allTLDBackends {
-		bs = append(bs, *tldBackend)
-	}
-
-	return bs
-}
-
-func errorsFanIn(errs []error, nBackends int) error {
-	nErrs := len(errs)
-	var counts = make(map[string]int)
-	switch {
-	case (nErrs == 0):
-		return nil
-	case (nErrs < nBackends):
-		return nil
-	case (nErrs > nBackends):
-		return errors.New("got more errors than there are backends. Probably something is broken")
-	default:
-		// everything failed, nErrs == nBackends
-		nNotNotFounds := 0
-		for _, e := range errs {
-			counts[e.Error()] += 1
-			if _, ok := e.(types.ErrNotFound); !ok {
-				nNotNotFounds += 1
-			}
-		}
-
-		nMajority := (nBackends + 1) / 2
-
-		if nNotNotFounds < nMajority {
-			return types.ErrNotFound(fmt.Sprintf(
-				"majority of backends returned not found. %d total errors, %d not found",
-				nErrs, nErrs-nNotNotFounds))
-		}
-		message := fmt.Sprintf("all backends failed with mixed errors: %+v", counts)
-		if len(message) > 300 {
-			message = message[:300]
-		}
-		return errors.New(message)
-	}
+	logger.Info("lb request served", zap.Int("http_code", http.StatusOK), zap.Duration("runtime_seconds", time.Since(t0)))
+	app.Metrics.Responses.WithLabelValues(strconv.Itoa(http.StatusOK), "lbcheck").Inc()
 }
